requirements.txt
GPUtil
Levenshtein
matplotlib
nltk
numpy
pandas
plotly
rouge
scipy
spacy
tensorflow-gpu
